refactor evaluation page and delete QA container
